pyproject.toml
[tool.poetry]
name = "transcript_summariser_crewai"
version = "0.1.0"
description = ""
authors = ["Gabriel Vailati <gabrielgvailati@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.28.8"
langchain-groq = "^0.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
